pyproject.toml
[project]
name = "frame"
version = "0.1.0"
description = ""
authors = [
    {name = "Moritz Enderle",email = "contact@moritzenderle.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0,<0.116.0",
    "uvicorn[standard]>=0.29.0,<0.32.0",
    "jinja2>=3.1.0,<4.0.0",
    "python-multipart>=0.0.9,<0.0.12",
    "pillow (>=12.0.0,<13.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "immich-python-sdk (>=1.134.0,<2.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
